Reformat DocumentExportPresenter
This commit changes how DocumentExportPresenter obtains its information
for a document. Previously the approach was to have conditional logic
around all the associations a document could have, this will no longer
work well as we're planning to strip and add data to each field in a bespoke
way which would generate a hugggggge conditional.

Adding more data in later commits.