Add history information into an exported document
This adds the history of change information into an exported document.
Within Whitehall this uses the "versions" terminology which gives clues
to the [paper_trail gem](https://github.com/paper-trail-gem/paper_trail)
origin of this log. I figure versions doesn't mean much to anyone so
"revision_history" is a better name to capture this data.
"history" was also considered but was dismissed as it could be confused
with 'history mode'.