Remove edition fields that are duplicated in translations
Instead of duplicating fields like 'title', 'summary' and 'body'
on both the edition root and its associated (primary) translation,
we now remove these fields from the edition root so that they only
exist on the translations, meaning the edition root now only houses
metadata.
This means a smaller payload of data to migrate, and a consistent
way of dealing with all languages.

As part of the test for this, I check that edition does still
retain some information that is not duplicated in the translations,
i.e. 'primary_locale', otherwise there is a risk that a future
change could make the 'edition' hash empty and the tests would still
pass.