Stop locked documents from appearing in CSV export
Trello - https://trello.com/c/qgeWWHYa/1150-create-card-to-exclude-locked-doc-from-csv-export

As it stands when a user exports a CSV of documents from whitehall this includes
locked Documents that are either in the process of or has finished migrating to
content publisher. This should not be the case as once it has been migrated,
information that Whitehall has on the document may now be out of date. This PR
ensures that locked documents are no longer included in the CSV.

We've manually tested this in integration and it successfully changes the URL at
the bottom of the search results and stops Editions from appearing in the CSV report

Co-authored-by: Bruce Bolt <bruce.bolt@digital.cabinet-office.gov.uk>