Include user details in document export
This includes a collection of users who have worked on a particular
document as a collection. This is done with the intention that all of
these users can be imported into Content Publisher via this script
rather than requesting their data from Signon.

The option to not use signon was chosen as this would require additional
endpoints adding to signon. Importing data between apps is unlikely to
risk much problems as when a user signs in their data will be re-synced
which should resolve any stale data problems.

In approaching this I considered either building the list of users at
the end by re-traversing the document or by having a collector variable
on the class that would record each time a user_id was encountered. I
decided to go with the former as I felt that this kept the code cleaner
with more focus on specific tasks.