Export variants of image sizes
When we import content from Whitehall, we need to ensure that we
transfer all the variants of image sizes too, in case they are
used by frontend applications. This is particularly true of image
thumbnails.
There is no easy way to query image variants from outside of
Whitehall, so this work is best done as part of the export.

Trello card: https://trello.com/c/XOJ64DPo/1151-export-variants-of-images-and-attachments