Export variants of attachments
When we import attachment content from Whitehall, we need to
ensure that we transfer all the variants of attachments too,
in case they are used by frontend applications. This is
particularly true of attachment thumbnails.

As well as the URL of the attachment variant, I export the
content type of the attachment variant as these can in theory
vary quite a bit. For instance, a PDF attachment may have a
PNG thumbnail variant, or perhaps another PDF variant that is
shorter/smaller, for example. It is safer to retain all of this
information.

Trello card: https://trello.com/c/XOJ64DPo/1151-export-variants-of-images-and-attachments