Guard against database failures in the HealthcheckController
The overdue URL is accessed by the govuk_uptime_collector, and this
happens on Staging while the database is being synced from Production.

As the database is out of action when being restored, there are errors
getting to Sentry relating to the query here failing. Rather than
letting these errors get to Sentry, catch the exception and log the
message.