Do not overwrite CSS class for buttons
In 3714c054f0, a change was made that replaced replaced the class on all
links to `govuk-link`.  Since buttons are styled links (with a `button`
class), this meant that the buttons were being styled as links when
being published.

This change retains the existing class(es) for buttons and adds tests to
ensure this change will get caught in future.