Replace internal links in documents that link to a migrated document
Whitehall allows publishers to link to other Whitehall documents using
an internal URL (rather than the public one).  On migration of a
document out of Whitehall, links to this document will no resolve.
Therefore, we need to replace internal links to this document in other
documents with the public URL.