Make some editions not default political by ministerial association
For a certian set of document types (and some specific types of
publications), don't by default mark them as political if they're
associated with a minister.

This should be a better default behaviour. To be clear, it'll still be
possible to mark the content as political, it just won't be done by
default.