Stop sending role appointment links
This changes the links we send to the Publishing API to make use of the
new reverse links that have been added.

See https://github.com/alphagov/publishing-api/pull/1645 for more
details.

The main benefit to doing this is to avoid having to maintain four sets
of links for each role appointment (person -> appointment, appointment ->
person, role -> appointment, appointment -> role) and instead we only
have two links (appointment -> person, appointment -> role) and we can
use the Publishing API to correctly present the reverse links back to
the appointment on the person and the role.

This means we don't need to worry about the links ending up out of sync
with each other which in practice would've meant that we had added some
`after_save` hooks on role appointments which republish the person and
role to the Publishing API.