Support Brexit CTA govspeak conversion
There is a need to provide a Brexit CTA which can be easily reused across
multiple documents and removes the need for content designers to update its
content in multiple places when the messaging needs to change. This implements
a pre-processing Govspeak helper which looks for occurrences of `$BrexitCTA` in
a document's content and replaces it with a chunk of CTA markdown with the
required messaging. This markdown will then be converted to HTML at the same
time as the rest of the Govspeak that has been provided.

We've decided to replace `$BrexitCTA` with markdown instead of hardcoded HTML
as this should ensure that it remains consistent with other CTAs and should
avoid any potential surprises in the event that Govspeak's behaviour ever
changes for any of the elements we use.