Restrict access to historic editions
Currently we limit creating a new edition of a historic document to
managing and GDS editors. However, an edition that was drafted before
the document became historic would not be included in this restriction
meaning that it could be edited and published by any editor.

Departmental publishers shouldn't be able to interact in any way with
content that is historic to preserve it as a record of the government
of the time.

The reason why this wasn't working before is that we were only
restricting access to these actions if the `action == :modify` which
isn't a real action. Instead I've limited all actions except for `:see`
and added the flash message redirect to the workflow controller so users
understand why the operation failed.