Work around an odd issue with atom feeds
Somehow, with recent versions of the govuk_publishing_components,
`.try(:document)` is calling a function defined in the
_document_list.html.erb partial in the govuk_publishing_components...

I don't quite understand why, so switch the code around to avoid
calling try on objects that don't respond to .document. In the case of
english responses, the record will be a RummagerDocumentPresenter,
which doesn't respond to .document, so just return it's id. In the
case of non english responses, the record is a presented edition,
which does respond to .document.