Update to verison 5.3.0 and add extends FriendlyId to all parent and sub classes
Running the test suite, multiple failures occur indicating that the 'friendly'
method is being called on an empty ActiveRecord::Relation array - leading to
a NoMethodError.

Re-running a failing test with backtrace turned on points towards
the following line (which was added as part of the FriendlyId
v5.2.5 release)...

relation = self.class.base_class.unscoped.friendly

See commit:  https://github.com/norman/friendly_id/commit/d67712ea62b7c6aebcf42a38749d77ca5524eb42

This commit seems to enforce the FriendlyId project's view that
when using Single Table Inheritance (STI) - which the current Whitehall
implementation does - "you must extend FriendlyId in all classes that
participate in STI".  A viewpoint that supports the change being a patch
version change (from 5.2.4 to 5.2.5) and not a major or minor version
change.  Other than a basic commit comment, documentation surrounding
the commit, and the reasoning behind it, is limited.

Previously, the intent within Whitehall is that only HtmlAttachment objects
are 'friendly', and other sub-classes of Attachment are not.  Something,
that does not correspond with the projects viewpoint and would most likely
be considered an incorrect use of the gem.

By extending FriendlyId in all STI classes it means that the Whitehall
implementation now supports the FriendlyId viewpoint, and so can now be
considered a 'correct' implementation.  Additionally, moving the
'friendly_id :title...' line up into the parent class (Attachment) means
all sub-classes now have the same 'slug' definition - based on the common
'title' attribute and is therefore consistent for all Attachment types.

Re-running the test suite no longer causes failures.

As Whitehall now uses FriendlyId in the intended manner, and as long
as that viewpoint is held by the maintainers of the project, we
should be able to continue using FriendlyId without having to freeze
our version at 5.2.4, and be confident that the implementation supports
the intended use of the gem.  The hope is that this leads to fewer,
and easier to manage, dependabot PR's in future.