Add a scope for filtering on link report links
This allows us to filter on link report links which can be deleted. The
link table has exceeded 10 GB and is the biggest Whitehall table so we
need to find a way of reducing the size. We've decided that links with a
status of `ok` which were created over 6 months ago can be deleted.