Fix flakey test
This test was flakey for a few reasons, one being for both the
organisations associated with the edition they both had a lead_ordering
of 1 which isn't a realistic situation and led to the query 50% of the
time returning an empty documents array as the wrong lead organisation
was being compared due to the lead_ording ASC LIMIT 1 part of the query.

The second layer of flakeyness was comparing the order of the lead
organisations in the response as these are not ordered in the actual
query so ordering was random and so broke occasionally complaining
about them being in the wrong order.

I've also made other minor test improvements with wording of the tests
and what we should actually be expecting instead of looking at the full
response.

Trello:
https://trello.com/c/5e7mzxrV/1327-check-logic-for-whitehall-lead-orgs-and-fix-flaky-test