Removes lead orgs for latest edition code
These organisation content-ids are not used by content-publisher's
import script and are superfluous. We already have logic making sure
that the documents are returned only when the 'primary org' (lowest
numerical lead_ordering) is matched against the passed in organisation
so including these lead orgs in the document info stage is pointless.