Fix private method delegation warning
Using def_delegators on the translatable methods of a globalize enhanced
ActiveRecord model is at risk of causing warnings. This is because, as
far as I can tell, a reliance on meta programming and method missing for
method existence. This seems to kick in when the translatable record is
updated (presumably as it may need to reconfigure the methods of the
model). The Forwardable module tries to determine whether a
method exists publicly or not before calling otherwise it raises a
warning, so any dynamically handled methods get flagged for warnings.

For example:

```
/var/lib/jenkins/workspace/whitehall_master-QR64ZCMXGMHM677MVMNGGITR273AGV2LXFQYVZ3J33DFGKL556ZQ/app/presenters/publishing_api/contact_presenter.rb:26: warning: PublishingApi::ContactPresenter#title at /usr/lib/rbenv/versions/2.6.5/lib/ruby/2.6.0/forwardable.rb:158 forwarding to private method Contact#title
```

This problem does not occur when using the ActiveSupport provided
delegate method.