Fix tests that have never asserted
These tests previously didn't assert and just returned true or false as
to whether the link that was expected was present or not. Unsurprisingly
it then turned out that once an assertion was added these tests failed
as they either a) never worked or b) got left behind by the app
evolving.