Handle Brakeman warning: Parameters should be whitelisted for mass assignment
Brakeman now doesn't allow implicitly permitting all params with a bang
and wants us to be explicit with every parameter we're permitting. This
makes sense but unfortunately whitehall doesn't really follow this
protocol in quite a few places, a way around permitting every param is
slicing the params and then permitting the following object.