Ignore friendly for a file and external attachment
Friendly by default uses the slug of a record to generate a route param
when generating a link, if a slug isn't present then the
default rails behaviour is used which uses the id.

For example an attachment with a slug field set to 'number-1':
Without friendly
E.g link_to "Attachment", @attachment
-> /attachment/1
With friendly
E.g link_to "attachment", @attachment
-> /attachment/number-1

We need to overide this behaviour as we were facing a production error
which meant the slugs between attachments were the same leading to the
wrong attachment being opened when visiting the edit page for the other
attachment.

The solution is now to use the attachment id's for routing like was
implemented previously (before this gem bump) for non HTML attachments /
foreign HTML attachments. We can do this via returning false for
`should_generate_new_friendly_id?` which prevents slugs being created
for certain attachments that were never meant to have them in the first
place... problem solved!

Also the friendly routes need to be set to false, this is to stop
friendly from using slugs as part of route generation for file and
external attachments. This shouldn't happen anyway as there shouldn't be
slugs on these attachments but due to the data sync from production to
integration some slugs obtain a value. Setting routes to false stops
these being used.