Set all non HTML attachments slugs to nil
Non HTML attachments shouldn't have slugs but this was allowed to
happen in the initial bump of this gem. At the time we thought this
wouldn't make a difference but a bug was introduced so this is a
migration to revert all slugs to nil given they're not html
attachments.