Add params for publications and consultations controllers
We now use Whitehall::DocumentFilter::CleanedParams to permit the
relevant params, publications_controller already uses this to convert
the params to something finder-frontend can understand so we can be
confident this covers the range of parameters being sent to these
controllers.

Originally I updated the params manually until all the tests passed but
this relied on the tests covering 100% of the parameter in use which
seemed unlikely.

These finders are pretty much deprecated in favour of those in
finder-frontend and only exist for foreign translations of documents.