Export contacts for all documents
Previously contacts were only exported for documents that had been
force_published, published or unwithdrawn. This was because we were
relying on an edition association - depended_upon_contacts which is only
populated when a document transitions to one of the states stated above.

Consequently draft editions never had any contacts information exported
and so when importing one in content-publisher an error was thrown
regarding missing contacts. The change in this commit now looks through
the body of text of the edition to pull out the relevant contacts.

Trello:
https://trello.com/c/OzdqkkW2/1455-fix-documents-that-failed-the-import-with-an-array-error