Remove link between About pages and publications
About pages contain a CorporateInformationGroups section that's
rendered conditionally on the presence of "transparency pages"
(FOI release, corporate reports or transparency data), which
doesn't make any sense because there are other links that could
still be rendered without such pages existing (such as "Jobs).

This is a blocker for moving these document types to Content
Publisher, as we don't want anything to be inadvertently
republished.