Change ordering of callbacks in take part page model
In carrierwave 2.x, the storing of images to their remote location has
been moved from after_save to after_commit (but only on create and update,
so mimicking the behaviour of after_save) [1].  These are executed after
after_save.

Whitehall is publishing the document in an after_commit callback which
appears to only upload to remote storage after the document has
published.  We therefore had a failing test case where the image had not
been uploaded at the point of the document being uploaded.

Rails executes after_commits callbacks in reverse order [2], so moving the
include of PublishesToPublishingApi to before the carrierwave uploader.

[1]: https://github.com/carrierwaveuploader/carrierwave/commit/665f2254a2c70385b67942b13a4710dc6826c42d
[2]: https://github.com/rails/rails/issues/20911