Change cached image used to resize images
We received a Zendesk ticket [1] about a loss of quality in thumbnail images
for a Person. While the uploaded image looks to be processed correctly in
Whitehall, the thumbnail representation showed a significant degradation in
quality. After investigation we found an issue raised on the Carrierwave gem
[2] which suggested that repeatedly recreating versions of an image is causing
degradation in image quality for the resulting variants. In our Image Uploader
we use the `:from_version` option to get different image variants, which uses a
cached version of the file instead of the original version, presumably since
according to the README for Carrierwave it "potentially results in faster
processing" [3]. Our existing approach to using the `:from_version` option is
to use the cached file for the variant that is the next size up from the
variant being created. This repeated resizing of images that have already been
processed seems to be the cause of the image quality degradation; switching all
variants to use the cached file for the largest variant seems to fix the issue.

[1] https://govuk.zendesk.com/agent/tickets/3923306
[2] https://github.com/carrierwaveuploader/carrierwave/issues/2436
[3] https://github.com/carrierwaveuploader/carrierwave#create-versions-from-existing-versions