Fix find_asset bug
Rails.application.assets resolves to nil which causes a no method error
when find_asset is run. I'm unsure what exactly changed to cause
Rails.application.assets to be set to nil, apparently this was a common
issue when upgrading to sprockets-rails version 3. It also could be
related to a recent bump of the carrier-wave gem
(https://github.com/alphagov/whitehall/pull/5253) but it's not obvious
what would cause this to happen.

This commits implements a workaround for find_asset which seems to work
as expected.