Use a more accurate timestamp for withdrawals
Previously we used the 'updated_at' time of a withdrawn edition to
represent the time it was withdrawn. We have an issue where a bulk
update to a bunch of withdrawn editions advanced the timestamp to
the time the update was performed, instead of preserving the time.

This fixes the issue by using the 'created_at' time of the actual
unpublishing object associated with the original withdrawal, noting
that we do not replace this original withdrawal when changing the
unpublishing explanation (it's edited in-place).

irb(main):002:0> Edition.find(1021159).unpublishing
=> #<Unpublishing id: 45163, edition_id: 1021159, unpublishing_reason_id: 5, explanation: "EU rules continue...

irb(main):003:0> Edition.find(1021159).unpublishing
=> #<Unpublishing id: 45163, edition_id: 1021159, unpublishing_reason_id: 5, explanation: "Ben test"...