Send attachment IDs in featured attachments list
Frontend apps will render these by looking up the ID in the
attachments list and chucking the metadata into a component.

This change looks a bit strange, because all the attachments IDs get
included, so a reasonable question is "what's the point? why not just
render 'attachments' as featured attachments if it's a publication or
consultation?"

This was addressed in the RFC but, in summary, there's no reason why a
format or document couldn't have both featured and non-featured
attachments, and there's no guarantee that the current situation will
remain the case.  So by adopting this approach, we don't need to
change things in the future.  You could say "YAGNI", but it's not like
it adds a lot of complication.