Add task to republish publications synchronously
This is to populate the new featured_attachments field.  The commit
can be reverted after the task has been run.

Republishing synchronously (in a single thread) is far slower than
asynchronously.  The upside is that it means the publishing load is
evenly spread out over a long period of time, so the publishing-api
and downstream services aren't overwhelmed by ~160,000 docs being
added to a queue all in one go.

Based on how long the consultations took, I think this will take about
16 hours to run