Fix buggy '.' in command paper number prefix validation
The validation uses a regular expression that re-interprets the
'.' in these prefixes to mean 'any character'. The end result is
that you can enter prefixes like 'CM' and 'Cmd', or even 'CmndZ',
none of which are valid prefixes. I've now made the regex escape
any special values, so that it is treated as an explicit period.
I considered altering VALID_COMMAND_PAPER_NUMBER_PREFIXES itself,
but as this is used to populate the validation message I didn't
have many other options.

I've also extended the tests to protect against double-spacing
and uppercasing. Double spaces were allowed as a result of the
'.' in the regex: this can lead to inconsistencies such as
"C  123" being valid as it matches "C. 123".

Uppercasing was also allowed as a result of the buggy regex;
"CM 123" would not be considered invalid against "Cm. 123" as
perhaps expected, but instead would be valid against "C. 123".