Write rake task to fix existing 'command_paper_number' values
Have added to the 'data hygiene' namespace, since that seems
appropriate, and followed the same convention used previously:
https://github.com/alphagov/whitehall/pull/4610

There is a 'dry run' and 'real run' option for the rake task, so
that we can see what it will do before we commit to it.
I've had to refactor the AttachmentValidator to decouple the
validation business logic from the Rails validation modelling, so
that we can call the validator from the AttachmentAttributeUpdater.