Use whitehalls email signup for world locations
If an atom_feed_url matches a world feed then forward the request to the
email_signups controller, previously this returned an
email-alert-frontend url but after a bug was caught this is now being
updated to use whitehall email_signup flow.

A rescue has been added to catch if the atom_feed_url is invalid e.g
integer, nil etc, if the url is invalid we'll redirect users to gov.uk's
homepage.