Export alternative_path instead of alternative_url
When checking an imported unpublishing alternative URL against the alternative
path present in Publishing API, we have found that Whitehall sends the path,
not the URL, to Publishing API. As such, the check was failing. This changes
the Document Export Presenter to send the unpublishing alternative path instead
so that these conflicts can be avoided.