Add support for 'CP' prefix
As of 2019, Command Papers are prefixed 'CP'. Interestingly,
unlike the other prefixes, 'CP' should not be followed by a
period.

I've taken the opportunity to reorder 'CP' to the beginning of the
array of allowed prefixes, as suggested by the content designer,
since we expect this to become more commonly used over time.