Add support for Roman numeral suffix on command paper numbers
This involved adding a `(-[IV]+)?` to the regex, but the tests
would continue to pass (when they shouldn't) since the `?` makes
that particular suffix optional. So I had to add a `\z` to the
end of the regex, to denote that nothing else is allowed after
the number except this specific optional suffix.

Unfortunately, when adding the `\z`, test/unit/attachment_test.rb
would fail with seemingly unrelated errors:

NameError: undefined local variable or method `not_deleted' for #<ActiveRecord::Relation []>
    app/models/attachable.rb:6:in `block (2 levels) in <module:Attachable>'
    app/models/file_attachment.rb:74:in `filename_is_unique'
    test/unit/attachment_test.rb:53:in `block (2 levels) in <class:AttachmentTest>'

It seems the built-in `validates` helper has its limitations -
and the regex was now growing complicated enough to warrant
splitting out into multiple lines - so I took the decision to move
the validation into the AttachmentValidator, following the
convention of the HOC paper number validation.

In the next commit I'll move all command paper number validation
tests into the same suite for consistency.