Use attachment ID as publishing-api ID
Using filename for file attachments isn't safe, as a document could
have two attachments with the same filename.  This results in
publishing-api rejecting the document when whitrehall tries to publish
it, as there's a constraint that the IDs of featured attachments are
unique.

We've seen this in integration, because the data sync renames a lot of
attachments to "redacted.pdf", but it could also happen in production.
We've just not seen that yet because people tend to use different
filenames for all the attachments on a single document.

Using "{id}-{filename}" would work but, in that case, why not just use
the id directly?

There's no need to republish docs after this, as the old IDs are still
valid.