Allow unpublished_at date to be passed directly
This makes the `PublishingApiWithdrawalWorker` class accept an extra
argument for the `unpublished_at` date. This is useful because currently
this comes from the `updated_at` of the edition which can change even
after a document is withdrawn, particularly after a data migration.

The argument is optional to support any workers that might still be
running during the migration.

Unfortunately, by having to put the argument at the end, `allow_draft`
effectively becomes a requried argument since all future calls to this
worker will pass this argument in. Once this is deployed, I'll make both
of these arguments required.