Pass unpublished_at date when withdrawing
This changes the withdraw process to pass the created_at date of the
unpublishing object when withdrawing. This date is unlikely to change,
whereas the current date (the `updated_at` of the edition) is able to
change which means that when re-publishing a document could end up with
the wrong date.

This fixes a potential problem where the withdrawn date might end up
being inaccurate if an edition changes for other reasons. In particular,
this is most likely to happen when writing a data migration.