Scope attachment validations to user submissions
Previously we applied attachment validations for user and internal
actions, which resulted in a bug where a new draft of an edition
would 'shed' any invalid attachments (since they wouldn't save).

In general, we should only apply 'semantic' (i.e. doesn't lead to
a technical error) validations at the point of entry, and not for
other operations where the user is not at a fault.

This updates the Attachment model to scope the AttachmentValidator so
that it's only part of the model validations when a user creates or
updates an attachment. (Or similarly via bulk upload.)