Fix the only_broken_links Edition method
I think the purpose of this is to limit the scope of the query to
editions with broken links. There can be more than one report for a
single edition though, which makes sense. This query however makes no
sense, as it doesn't just look at the latest report, but any
report. So it's more like asking what editions have ever had a broken
link!

I was looking in to this because of reports of it not working due to
timeouts.

This commit changes the query to join on just the latest reports for
each edition, and then only select editions where a broken link exists
for that latest report. I don't know much about MySQL, let alone
ActiveRecord, hence the big blob of SQL.