Delete LinkCheckerApiReport's without links
I don't think having reports without any associated links is useful,
so go ahead and delete these reports as well as deleting the related
links.

The intent here is to potentially help with the performance of the
Edition.only_broken_links method, by not having so much data to search
through.