Move author notifications to a Sidekiq worker
This should help to avoid issues when sending author notifications,
which currently can happen in the ScheduledPublishingWorker.

If the sending of the email fails, due to rate limiting for example,
previously the ScheduledPublishingWorker would retry, but then succeed
immediately because the content has been published. Because the author
notifications aren't part of the transaction involved in the scheduled
publishing, it's appropriate to use a Sidekiq worker, where the
retrying can be handled separately.