Silence RuboCop errors about unique indices
Although uniqueness constraints should be defined at the DB level, as
explained by the docs for this cop, this would involve dropping and
recreating the associated indices, which is a relatively risky change in
the context of a linting fix. We can still benefit from this cop, but
I don't think we need to be applying it retrospectively.