Remove transitive asset replacement logic
This is handled in Asset Manager, and has been for a while.  This code
as it is causes problems with redirect chains.  Say you have two live
assets:

    Asset 1 -> Asset 2

(where `->` means "is replaced by")

Visiting Asset 1 redirects you to Asset 2, good stuff.

Then you add a third, draft, asset, as a replacement for Asset 2.
This transitive closure logic kicks in, and we end up with:

    Asset 1 -> Asset 3
    Asset 2 -> Asset 3

As Asset 3 is a draft, neither Asset 1 nor Asset 2 redirect to it.
But the redirect from Asset 1 to Asset 2 is now broken!

Asset Manager handles this properly, so stop trying to be clever and
just let it do its thing.